mkdtimg: Bundle upstream python rewrite as submodule
The old C version is now deprecated, so we don't have to bother with
integrating that into the build system. That makes it quite easy for us
to bundle upstream libufdt as a submodule in dtc/libufdt (just beside
dtc/libfdt), and therefore not require exotic out of tree build
dependencies.

Signed-off-by: KyuoFoxHuyu <88403766@qq.com>
Signed-off-by: RuRuTiaSaMa <1009087450@qq.com>