qcacld-3.0: Ensure proper seg0 seg1 bandwidth values
Some APs advertise wrong seg0, seg1, and BW values
and the driver copies that info directly and sets it
to the FW, which can lead to a crash in HAL layers
due to the wrong configuration.

Fix is to ensure that only correct combination
allowed by regulatory can be passed onto the FW
in the vdev start command.

Change-Id: I452438f385d6f067812551071b2ea87b99b9eb01
CRs-Fixed: 2698732
Signed-off-by: RuRuTiaSaMa <1009087450@qq.com>