sched/uclamp: Use wrappers for *_{read, write, show} functions
In commit:
    sched/uclamp: Move all tunables to cpusets,

We directly modified and exported *_{read, write, show} functions
to cpuset code.This makes the resulting code quite inconsistent with
the upstream version.

We now use wrappers to reflect the original function, and to preserve
the code in accordance with upstream/mainline.

This is a purely cosmetic change, and no new behaviour should be
exhibited on applying this commit.

Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
(cherry picked from commit 5d3e10eba8ac4b29fccf7b500d09c6adf8becca0)
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: RuRuTiaSaMa <1009087450@qq.com>