drivers/cpufreq: cpufreq_performance: Force set CPU frequencies to max/min according to profile set
The Performance governor forces the frequencies of the set cluster to max which does increase power consumption if one were to not be using an optimized efficient frequency table hence guarding it with kprofiles would be the better option as freqs will be forced to max only in level 0, 2 and 3. If kprofiles were to be set in level 1, the freqs will be forced to min.

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Signed-off-by: RuRuTiaSaMa <1009087450@qq.com>